+++ api/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Patch profiles now include a gear menu to set version overrides (or choose "All versions") per profile
 - Added Korean manager string localization https://github.com/Jman-Github/Universal-ReVanced-Manager/pull/42
 - Split APKs now save in Settings > Downloads as merged APks https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/41
+- Gave the GitHub PAT entry in Settings > Advanced the ability to be saved through the manager settings exports. This is a toggable feature and is not on by default
 
 
 # Bug fixes